backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
faiss-cpu
numpy
pydantic
starlette
